Extract Terraform CLI model
